pyinstaller fix for linux
and move the pyinstaller hooks out of the git root directory

The problem was that the file `mip/libraries/cbc-c-linux-x86-64.so` is not found.
Digging into pyinstaller hooks, we can see that:

- `collect_all()` calls `collect_dynamic_libs()` with mostly default args - see
[here](https://github.com/pyinstaller/pyinstaller/blob/919022fa471238de7fb832c81f6c1334b79604ab/PyInstaller/utils/hooks/__init__.py#L1144)
- `collect_dynamic_libs()` looks for libraries using the arg `search_patterns`
which defaults to `PY_DYLIB_PATTERNS` - see [function def starting here](https://github.com/pyinstaller/pyinstaller/blob/919022fa471238de7fb832c81f6c1334b79604ab/PyInstaller/utils/hooks/__init__.py#L723)
- `PY_DYLIB_PATTERNS` is defined [here](https://github.com/pyinstaller/pyinstaller/blob/919022fa471238de7fb832c81f6c1334b79604ab/PyInstaller/utils/hooks/__init__.py#L716)

``` python
PY_DYLIB_PATTERNS = [
    '*.dll',
    '*.dylib',
    'lib*.so',
]
```

So it finds:

```
├── cbc-c-darwin-x86-64.dylib
└── win64
    ├── cbc-c-windows-x86-64.dll
    ├── libbz2-1.dll
    ├── libgcc_s_seh-1.dll
    ├── libstdc++-6.dll
    └── libwinpthread-1.dll
```

but **fails** to find:

```
├── cbc-c-linux-x86-64.so
```

It **would** find it if it was `libcbc-c-linux-x86-64.so` - I'm not sure why PyInstaller uses `lib*.so` rather than `*.so`

So this change explicitly looks for `*.so` but deduplicates with
anything already found.
