+++ .github/workflows/build-executables.yml
@@ -32,7 +32,7 @@ jobs:
           uv sync
       - name: "Build windows package"
         run: |
-          uv run python -m eel script.py web -n strat-select-win.exe --additional-hooks-dir=. --onefile
+          uv run python -m eel script.py web -n strat-select-win.exe --additional-hooks-dir=./pyinstallerhooks/ --onefile
           mkdir -p target/release
           mv dist/strat-select-win.exe target/release
       # copy script exe to build artifacts so we can test it
@@ -65,7 +65,7 @@ jobs:
           uv sync
       - name: "Build MacOS package"
         run: |
-          uv run python -m eel script.py web -n strat-select-macos --additional-hooks-dir=. --onefile
+          uv run python -m eel script.py web -n strat-select-macos --additional-hooks-dir=./pyinstallerhooks/ --onefile
           mkdir -p target/release
           mv dist/strat-select-macos target/release
       # copy script app to build artifacts so we can test it
@@ -98,7 +98,7 @@ jobs:
           uv sync
       - name: "Build Linux package"
         run: |
-          uv run python -m eel script.py web -n strat-select-linux --additional-hooks-dir=. --onefile
+          uv run python -m eel script.py web -n strat-select-linux --additional-hooks-dir=./pyinstallerhooks/ --onefile
           mkdir -p target/release
           mv dist/strat-select-linux target/release
       # copy script app to build artifacts so we can test it
